Add packages registry.json file
This file will contain the list of packages that are available to the
sammy app and any other programs.

To add a new library to sammy, one will need to add a line to this json
file where the "key" is the package name. That package name should always
match the repo name unless that name is already taken. The "value"
should be the clone URL of the package.